[community] Update copyright information. NFCI
In the contribution guidelines [1] we state that Arm do not require
copyright assignment for any kind of contributions. The update is
needed to facilitate the update of the copyright infos for new
contributions.

We have left the pre open source reelase copyright statements, and
added the names of the people that have contributed so far since the
first opens source release [2].

We have also updated the License and Copyright sections of the
specifications to make them self contained in the document itself,
without reference to files that might not be shipped with the PDF file
of the specs.

Note: there are no functional changes in term of license.

[1] https://github.com/ARM-software/acle/blob/main/CONTRIBUTING.md#contribution-agreement
[2] https://github.com/ARM-software/acle/releases/tag/r2021-07-02